Update draft-ietf-ccamp-bwa-topo-yang.txt
row 19: Should the name of the model really include "latest"?
row108-118: example "2" is in principle the same as example "1", and the text is a copy from an earlier section in this chapter. I suggest that we remove it.
row209: extra empty row. Remove!?
row206, 212, 213, 283: replaced matrix with table to align with group name "link-bw-availability-table"
row229: date and description needs to be updated.
row258: empty. remove!?
row260-261: Updated the description
row314: Updated text
row317: Not required. Can be removed.
row426: Replaced the tree with a new small example tree